chore: rename hooks.ts into redux.ts
