+++ src/shared/ui/ListItem/ListItem.tsx
@@ -5,7 +5,7 @@ import React, {useState} from "react";
 import edit from "@/shared/icons/edit.svg";
 import remove from "@/shared/icons/remove.svg";
 import styles from "./ListItem.module.css"
-import {useAppSelector} from "@/shared/hooks/hooks";
+import {useAppSelector} from "@/shared/hooks/redux";
 import {ModalEditItem} from "@/features/ModalEditItem";
 import {ModalRemoveItem} from "@/features/ModalRemoveItem";
 import {Item} from "@/shared/model/item";
